some more commands moved to modules
