+++ src-tauri/src/gui/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod utils_commands;
+pub mod prequisites;
+pub mod installation;
+pub mod settings;
