+++ src-tauri/src/gui/mod.rs
@@ -21,463 +21,19 @@ use std::{
 use tauri::{AppHandle, Manager}; // dep: fork = "0.1"
 mod app_state;
 mod ui;
-
+mod commands;
+pub mod utils;
 
 use app_state::{AppState, get_locked_settings, get_settings_non_blocking, update_settings};
-use ui::{send_message,send_install_progress_message, send_simple_setup_message,send_tools_message, emit_to_fe, ProgressBar};
+use ui::{send_message,send_install_progress_message, send_simple_setup_message,send_tools_message, ProgressBar};
+use commands::{utils_commands::*, prequisites::*, installation::*, settings::*};
+use utils::is_path_empty_or_nonexistent;
 
 // Learn more about Tauri commands at https://tauri.app/v1/guides/features/command
 #[tauri::command]
 fn greet(name: &str) -> String {
     log::info!("Greet called with name: {}", name);
     format!("Hello, {}! You've been greeted from Rust!", name)
-}
-
-#[tauri::command]
-fn get_settings(app_handle: tauri::AppHandle) -> Settings {
-    get_locked_settings(&app_handle).unwrap_or_default()
-}
-
-#[tauri::command]
-fn get_prequisites() -> Vec<&'static str> {
-    idf_im_lib::system_dependencies::get_prequisites()
-}
-
-#[tauri::command]
-fn get_operating_system() -> String {
-    std::env::consts::OS.to_string()
-}
-// ----
-#[tauri::command]
-fn install_prerequisites(app_handle: AppHandle) -> bool {
-    let unsatisfied_prerequisites = idf_im_lib::system_dependencies::check_prerequisites()
-        .unwrap()
-        .into_iter()
-        .map(|p| p.to_string())
-        .collect();
-    match idf_im_lib::system_dependencies::install_prerequisites(unsatisfied_prerequisites) {
-        Ok(_) => true,
-        Err(err) => {
-            send_message(
-                &app_handle,
-                format!("Error installing prerequisites: {}", err),
-                "error".to_string(),
-            );
-            log::error!("Error installing prerequisites: {}", err);
-            false
-        }
-    }
-}
-
-#[tauri::command]
-fn check_prequisites(app_handle: AppHandle) -> Vec<String> {
-    match idf_im_lib::system_dependencies::check_prerequisites() {
-        Ok(prerequisites) => {
-            if prerequisites.is_empty() {
-                vec![]
-            } else {
-                prerequisites.into_iter().map(|p| p.to_string()).collect()
-            }
-        }
-        Err(err) => {
-            send_message(
-                &app_handle,
-                format!("Error checking prerequisites: {}", err),
-                "error".to_string(),
-            );
-            log::error!("Error checking prerequisites: {}", err);
-            vec![]
-        }
-    }
-}
-
-#[tauri::command]
-fn python_sanity_check(app_handle: AppHandle, python: Option<&str>) -> bool {
-    let outpusts = idf_im_lib::python_utils::python_sanity_check(python);
-    let mut all_ok = true;
-    for output in outpusts {
-        match output {
-            Ok(_) => {}
-            Err(err) => {
-                all_ok = false;
-                send_message(
-                    &app_handle,
-                    format!("Python sanity check failed: {}", err),
-                    "warning".to_string(),
-                );
-                log::warn!("{:?}", err)
-            }
-        }
-    }
-    all_ok
-}
-
-#[tauri::command]
-fn get_logs_folder(app_handle: AppHandle) -> PathBuf {
-    match idf_im_lib::get_log_directory() {
-        Some(folder) => folder,
-        None => {
-            send_message(
-                &app_handle,
-                "Error getting log folder".to_string(),
-                "error".to_string(),
-            );
-            log::error!("Error getting log folder"); //TODO: emit message
-            PathBuf::new()
-        }
-    }
-}
-
-#[tauri::command]
-fn python_install(app_handle: AppHandle) -> bool {
-    match idf_im_lib::system_dependencies::install_prerequisites(vec!["python".to_string()]) {
-        Ok(_) => true,
-        Err(err) => {
-            send_message(
-                &app_handle,
-                format!("Error installing python: {}", err),
-                "error".to_string(),
-            );
-            log::error!("Error installing python: {}", err); //TODO: emit message
-            false
-        }
-    }
-}
-#[tauri::command]
-async fn get_available_targets(app_handle: AppHandle) -> Vec<Value> {
-    let app_state = app_handle.state::<AppState>();
-    // Clone the settings to avoid holding the MutexGuard across await points
-    let settings = {
-        let guard = app_state
-            .settings
-            .lock()
-            .map_err(|_| {
-                send_message(
-                    &app_handle,
-                    "Failed to obtain lock on AppState. Please retry the last action later."
-                        .to_string(),
-                    "error".to_string(),
-                )
-            })
-            .expect("Failed to lock settings");
-        guard.clone()
-    };
-    let targets = settings.target.clone().unwrap_or_default();
-    let available_targets = idf_im_lib::idf_versions::get_avalible_targets()
-        .await
-        .unwrap();
-    // available_targets.insert(0, "all".to_string());
-    available_targets
-        .into_iter()
-        .map(|t| {
-            json!({
-              "name": t,
-              "selected": targets.contains(&t),
-            })
-        })
-        .collect()
-}
-
-#[tauri::command]
-fn set_targets(app_handle: AppHandle, targets: Vec<String>) -> Result<(), String> {
-    info!("Setting targets: {:?}", targets);
-    update_settings(&app_handle, |settings| {
-        settings.target = Some(targets);
-    })?;
-    send_message(
-        &app_handle,
-        "Targets updated successfully".to_string(),
-        "info".to_string(),
-    );
-    Ok(())
-}
-
-#[tauri::command]
-async fn get_idf_versions(app_handle: AppHandle) -> Vec<Value> {
-    let app_state = app_handle.state::<AppState>();
-    // Clone the settings to avoid holding the MutexGuard across await points
-    let settings = {
-        let guard = app_state
-            .settings
-            .lock()
-            .map_err(|_| {
-                send_message(
-                    &app_handle,
-                    "Failed to obtain lock on AppState. Please retry the last action later."
-                        .to_string(),
-                    "error".to_string(),
-                )
-            })
-            .expect("Failed to lock settings");
-        guard.clone()
-    };
-    let targets = settings.target.clone().unwrap_or_default();
-    let versions = settings.idf_versions.clone().unwrap_or_default();
-
-    let targets_vec: Vec<String> = targets.to_vec();
-    let mut available_versions = if targets_vec.contains(&"all".to_string()) {
-        idf_im_lib::idf_versions::get_idf_names().await
-    } else {
-        // todo: handle multiple targets
-        idf_im_lib::idf_versions::get_idf_name_by_target(&targets[0].to_string().to_lowercase())
-            .await
-    };
-    available_versions.push("master".to_string());
-    available_versions
-        .into_iter()
-        .map(|v| {
-            json!({
-              "name": v,
-              "selected": versions.contains(&v),
-            })
-        })
-        .collect()
-}
-
-#[tauri::command]
-fn set_versions(app_handle: AppHandle, versions: Vec<String>) -> Result<(), String> {
-    info!("Setting IDF versions: {:?}", versions);
-    update_settings(&app_handle, |settings| {
-        settings.idf_versions = Some(versions);
-    })?;
-
-    send_message(
-        &app_handle,
-        "IDF versions updated successfully".to_string(),
-        "info".to_string(),
-    );
-    Ok(())
-}
-
-#[tauri::command]
-fn get_idf_mirror_list(app_handle: AppHandle) -> Value {
-    let app_state = app_handle.state::<AppState>();
-    // Clone the settings to avoid holding the MutexGuard across await points
-    let settings = {
-        let guard = app_state
-            .settings
-            .lock()
-            .map_err(|_| {
-                send_message(
-                    &app_handle,
-                    "Failed to obtain lock on AppState. Please retry the last action later."
-                        .to_string(),
-                    "error".to_string(),
-                )
-            })
-            .expect("Failed to lock settings");
-        guard.clone()
-    };
-    let mirror = settings.idf_mirror.clone().unwrap_or_default();
-    let mut avalible_mirrors = idf_im_lib::get_idf_mirrors_list().to_vec();
-    if !avalible_mirrors.contains(&mirror.as_str()) {
-        let mut new_mirrors = vec![mirror.as_str()];
-        new_mirrors.extend(avalible_mirrors);
-        avalible_mirrors = new_mirrors;
-    }
-    json!({
-      "mirrors":avalible_mirrors,
-      "selected": mirror,
-    })
-}
-
-#[tauri::command]
-fn set_idf_mirror(app_handle: AppHandle, mirror: String) -> Result<(), String> {
-    info!("Setting IDF mirror: {}", mirror);
-    update_settings(&app_handle, |settings| {
-        settings.idf_mirror = Some(mirror);
-    })?;
-
-    send_message(
-        &app_handle,
-        "IDF mirror updated successfully".to_string(),
-        "info".to_string(),
-    );
-    Ok(())
-}
-
-#[tauri::command]
-fn get_tools_mirror_list(app_handle: AppHandle) -> Value {
-    let app_state = app_handle.state::<AppState>();
-    // Clone the settings to avoid holding the MutexGuard across await points
-    let settings = {
-        let guard = app_state
-            .settings
-            .lock()
-            .map_err(|_| {
-                send_message(
-                    &app_handle,
-                    "Failed to obtain lock on AppState. Please retry the last action later."
-                        .to_string(),
-                    "error".to_string(),
-                )
-            })
-            .expect("Failed to lock settings");
-        guard.clone()
-    };
-    let mirror = settings.mirror.clone().unwrap_or_default();
-    let mut avalible_mirrors = idf_im_lib::get_idf_tools_mirrors_list().to_vec();
-    if !avalible_mirrors.contains(&mirror.as_str()) {
-        let mut new_mirrors = vec![mirror.as_str()];
-        new_mirrors.extend(avalible_mirrors);
-        avalible_mirrors = new_mirrors;
-    }
-    json!({
-      "mirrors":avalible_mirrors,
-      "selected": mirror,
-    })
-}
-
-#[tauri::command]
-fn set_tools_mirror(app_handle: AppHandle, mirror: String) -> Result<(), String> {
-    info!("Setting tools mirror: {}", mirror);
-    update_settings(&app_handle, |settings| {
-        settings.mirror = Some(mirror);
-    })?;
-
-    send_message(
-        &app_handle,
-        "Tools mirror updated successfully".to_string(),
-        "info".to_string(),
-    );
-    Ok(())
-}
-
-#[tauri::command]
-fn get_installation_path(app_handle: AppHandle) -> String {
-    let app_state = app_handle.state::<AppState>();
-    // Clone the settings to avoid holding the MutexGuard across await points
-    let settings = {
-        let guard = app_state
-            .settings
-            .lock()
-            .map_err(|_| {
-                send_message(
-                    &app_handle,
-                    "Failed to obtain lock on AppState. Please retry the last action later."
-                        .to_string(),
-                    "error".to_string(),
-                )
-            })
-            .expect("Failed to lock settings");
-        guard.clone()
-    };
-    let path = settings.path.clone().unwrap_or_default();
-    path.to_str().unwrap().to_string()
-}
-
-#[tauri::command]
-fn set_installation_path(app_handle: AppHandle, path: String) -> Result<(), String> {
-    info!("Setting installation path: {}", path);
-    update_settings(&app_handle, |settings| {
-        settings.path = Some(PathBuf::from(path));
-    })?;
-
-    send_message(
-        &app_handle,
-        "Installation path updated successfully".to_string(),
-        "info".to_string(),
-    );
-    Ok(())
-}
-#[tauri::command]
-async fn is_path_empty_or_nonexistent(app_handle: AppHandle, path: String) -> bool {
-    let path = Path::new(&path);
-
-    // If path doesn't exist, return true
-    if !path.exists() {
-        return true;
-    }
-
-    // If path exists, check if it's a directory and if it's empty
-    if path.is_dir() {
-        match fs::read_dir(path) {
-            Ok(mut entries) => {
-                if entries.next().is_none() {
-                    //it's empty
-                    return true;
-                }
-                let settings = get_locked_settings(&app_handle).unwrap();
-
-                let vers = match &settings.idf_versions {
-                    Some(v) => v,
-                    None => {
-                        send_message(
-                            &app_handle,
-                            "No IDF versions selected. Please select at least one version to continue."
-                                .to_string(),
-                            "error".to_string(),
-                        );
-                        return false; // something is broken we don't have any versions selected
-                    }
-                };
-                for v in vers {
-                    let new_path = path.join(v);
-                    if new_path.exists() {
-                        return false;
-                    }
-                }
-                true
-            } // true if directory is empty
-            Err(_) => false, // return false if we can't read the directory
-        }
-    } else {
-        //path is file which is conflicting with the directory
-        false
-    }
-}
-
-#[tauri::command]
-fn load_settings(app_handle: AppHandle, path: &str) {
-    let mut file = File::open(path)
-        .map_err(|_| {
-            send_message(
-                &app_handle,
-                format!("Failed to open file: {}", path),
-                "warning".to_string(),
-            )
-        })
-        .expect("Failed to open file");
-    let mut contents = String::new();
-    file.read_to_string(&mut contents)
-        .map_err(|_| {
-            send_message(
-                &app_handle,
-                format!("Failed to read file: {}", path),
-                "warning".to_string(),
-            )
-        })
-        .expect("Failed to read file");
-    let settings: idf_im_lib::settings::Settings = toml::from_str(&contents)
-        .map_err(|_| {
-            send_message(
-                &app_handle,
-                format!("Failed to parse TOML: {}", path),
-                "warning".to_string(),
-            )
-        })
-        .expect("Failed to parse TOML");
-    log::debug!("settings {:?}", settings);
-    let app_state = app_handle.state::<AppState>();
-    let mut state_settings = app_state
-        .settings
-        .lock()
-        .map_err(|_| {
-            send_message(
-                &app_handle,
-                "Failed to obtain lock on AppState. Please retry the last action later."
-                    .to_string(),
-                "error".to_string(),
-            )
-        })
-        .expect("Failed to lock settings");
-    *state_settings = settings;
-
-    send_message(
-        &app_handle,
-        format!("Settings loaded from {}", path),
-        "info".to_string(),
-    );
 }
 
 fn prepare_installation_directories(
@@ -1198,28 +754,6 @@ async fn start_installation(app_handle: AppHandle) -> Result<(), String> {
 }
 
 #[tauri::command]
-fn quit_app(app_handle: tauri::AppHandle) {
-    app_handle.exit(0);
-}
-
-#[tauri::command]
-fn save_config(app_handle: tauri::AppHandle, path: String) {
-    let mut settings = match get_locked_settings(&app_handle) {
-        Ok(s) => s,
-        Err(_) => {
-            return send_message(
-                &app_handle,
-                "Instalation config can not be saved. Please try again later.".to_string(),
-                "error".to_string(),
-            )
-        }
-    };
-
-    settings.config_file_save_path = Some(PathBuf::from(path));
-    let _ = settings.save();
-}
-
-#[tauri::command]
 async fn start_simple_setup(app_handle: tauri::AppHandle) {
     let settings = get_locked_settings(&app_handle).unwrap();
     send_simple_setup_message(&app_handle, 1, "started".to_string());
@@ -1275,74 +809,6 @@ async fn start_simple_setup(app_handle: tauri::AppHandle) {
     send_simple_setup_message(&app_handle, 10, "Installing IDF".to_string());
     let _res = start_installation(app_handle.clone()).await;
 }
-
-#[tauri::command]
-fn show_in_folder(path: String) {
-    #[cfg(target_os = "windows")]
-    {
-        match Command::new("explorer")
-            .args(["/select,", &path]) // The comma after select is not a typo
-            .spawn()
-        {
-            Ok(_) => {}
-            Err(e) => {
-                error!("Failed to open folder with explorer: {}", e);
-            }
-        }
-    }
-
-    #[cfg(target_os = "linux")]
-    {
-        let path = if path.contains(",") {
-            // see https://gitlab.freedesktop.org/dbus/dbus/-/issues/76
-            match std::fs::metadata(&path).unwrap().is_dir() {
-                true => path,
-                false => {
-                    let mut path2 = PathBuf::from(path);
-                    path2.pop();
-                    path2.into_os_string().into_string().unwrap()
-                }
-            }
-        } else {
-            path
-        };
-
-        // Try using xdg-open first
-        if Command::new("xdg-open").arg(&path).spawn().is_err() {
-            // Fallback to dbus-send if xdg-open fails
-            let uri = format!("file://{}", path);
-            match Command::new("dbus-send")
-                .args([
-                    "--session",
-                    "--dest=org.freedesktop.FileManager1",
-                    "--type=method_call",
-                    "/org/freedesktop/FileManager1",
-                    "org.freedesktop.FileManager1.ShowItems",
-                    format!("array:string:\"{}\"", uri).as_str(),
-                    "string:\"\"",
-                ])
-                .spawn()
-            {
-                Ok(_) => {}
-                Err(e) => {
-                    error!("Failed to open file with dbus-send: {}", e);
-                }
-            }
-        }
-    }
-
-    #[cfg(target_os = "macos")]
-    {
-        match Command::new("open").args(["-R", &path]).spawn() {
-            Ok(_) => {}
-            Err(e) => {
-                error!("Failed to open file with open: {}", e);
-            }
-        }
-    }
-}
-
-use tauri::Emitter;
 
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
@@ -1411,7 +877,7 @@ pub fn run() {
             save_config,
             get_logs_folder,
             show_in_folder,
-            is_path_empty_or_nonexistent,
+            is_path_empty_or_nonexistent_command,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
