+++ src-tauri/src/gui/ui.rs
@@ -1,11 +1,11 @@
 use log::{debug, info};
 use serde_json::{json, Value};
-use tauri::AppHandle;
+use tauri::{AppHandle, Emitter}; // dep: fork = "0.1"
 
 
 /// Emits a message to the frontend
 pub fn emit_to_fe(app_handle: &AppHandle, event_name: &str, json_data: Value) {
-    debug!("emit_to_fe: {} {:?}", event_name, json_data);
+    debug!("emit_to_fe: {} {:?}", event_name, json_data); //TODO: remove debug
     let _ = app_handle.emit(event_name, json_data);
 }
